decouple pub & sub scripts so that they can be run in separate terminals
